roles wise page and items showing

+++ app/Http/Controllers/RolesController.php
@@ -9,6 +9,7 @@ use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Routing\Controller;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
@@ -16,6 +17,15 @@ use Spatie\Permission\Models\Role;
 class RolesController extends Controller
 {
     use ValidatesRequests;
+    public $user;
+
+    public function __construct()
+    {
+        $this->middleware(function($request,$next){
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -23,6 +33,9 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('role.view')) {
+            abort(403,'Sorry !! You are Unauthorized to view any role.');
+        }
         $roles = Role::SimplePaginate(10);
         return view('backend.roles.index',compact('roles'));
     }
@@ -34,6 +47,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403,'Sorry !! You are Unauthorized to create any role.');
+        }
         $permissions = Permission::all();
         $permission_groups = User::getPermissionGroup();
         return view('backend.roles.create',compact('permissions','permission_groups'));
@@ -47,6 +63,9 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403,'Sorry !! You are Unauthorized to create any role.');
+        }
         $this->validate($request, [
             'name' => "required|unique:roles,name",
         ],[
@@ -71,6 +90,9 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403,'Sorry !! You are Unauthorized to edit any role.');
+        }
         $role = Role::findById($id);
         $permissions = Permission::all();
         $permission_groups = User::getPermissionGroup();
@@ -86,6 +108,9 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403,'Sorry !! You are Unauthorized to delete any role.');
+        }
         $this->validate($request, [
             'name' => "required|unique:roles,name,$id",
         ],[
@@ -112,6 +137,9 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
+        if (is_null($this->user) || !$this->user->can('role.delete')) {
+            abort(403,'Unauthorized Access');
+        }
         $role = Role::findById($id);
         if(!is_null($role)){
             $role->delete();
